refactor: Move doQueryContext to db_inspect
Doing this because now db_inspect is the only one using it
